+++ src/gameentry.h
@@ -116,6 +116,7 @@ public:
     QString eSLastPlayed = "";
     QString eSKidGame = "";
     QString eSSortName = "";
+    bool isFolder = false;
 
     // AttractMode specific metadata for preservation
     // #Name;Title;Emulator;CloneOf;Year;Manufacturer;Category;Players;Rotation;Control;Status;DisplayCount;DisplayType;AltRomname;AltTitle;Extra;Buttons