+++ src/pegasus.cpp
@@ -375,8 +375,6 @@ void Pegasus::assembleList(QString &finalOutput,
     // Always make dotMod at least 1 or it will give "floating point exception"
     // when modulo
     int dotMod = gameEntries.length() * 0.1 + 1;
-    if (dotMod == 0)
-        dotMod = 1;
     for (auto &entry : gameEntries) {
         if (dots % dotMod == 0) {
             printf(".");