+++ src/retrobat.cpp
@@ -126,8 +126,6 @@ void RetroBat::assembleList(QString &finalOutput,
     // Always make dotMod at least 1 or it will give "floating point exception"
     // when modulo
     int dotMod = gameEntries.length() * 0.1 + 1;
-    if (dotMod == 0)
-        dotMod = 1;
     finalOutput.append("<?xml version=\"1.0\"?>\n<gameList>\n");
     for (auto &entry : gameEntries) {
         if (dots % dotMod == 0) {