+++ src/xmlreader.h
@@ -36,10 +36,10 @@ public:
     ~XmlReader();
     bool setFile(QString filename);
     QList<GameEntry> getEntries(QString inputFolder);
-    void addEntries(const QDomNodeList &nodes, QList<GameEntry> &gameEntries,
-                    const QString &inputFolder);
 
 private:
+    void addEntries(const QDomNodeList &nodes, QList<GameEntry> &gameEntries,
+                    const QString &inputFolder, bool isFolder = false);
     QString makeAbsolute(QString filePath, const QString &inputFolder);
 };